+++ repo/src/main/java/com/parashift/onlyoffice/OnlyofficeSettings.java
@@ -1,5 +1,6 @@
 package com.parashift.onlyoffice;
 
+import com.parashift.onlyoffice.constants.Formats;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,6 +27,7 @@ public class OnlyofficeSettings extends AbstractWebScript {
             Set<String> editableMimetypes = configManager.getCustomizableEditableSet();
             responseJson.put("editableMimetypes", editableMimetypes);
             responseJson.put("convertOriginal", configManager.getAsBoolean("convertOriginal", "false"));
+            responseJson.put("supportedFormats", Formats.getSupportedFormats());
 
             response.setContentType("application/json; charset=utf-8");
             response.setContentEncoding("UTF-8");